requirements.txt
pipdeptree
django
numpy
simpy
djangorestframework